% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freebird-package.R
\docType{package}
\name{freebird-package}
\alias{freebird-package}
\alias{_PACKAGE}
\title{freebird: Estimation and Inference for High Dimensional Mediation Analysis}
\description{
Estimation and Inference for High Dimensional Mediation Analysis.
}
\author{
\strong{Maintainer}: Ruixuan Zhou \email{rzhou14@illinois.edu} [copyright holder]

Authors:
\itemize{
  \item Dave Zhao \email{rzhou14@illinois.edu} [copyright holder]
}

}
