% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ihmm-package.R
\docType{package}
\name{ihmm-package}
\alias{ihmm}
\alias{ihmm-package}
\title{ihmm: Estimation and Inference for High Dimensional Median Analysis}
\description{
Amazing statistical computations are being undertaken.
}
\author{
\strong{Maintainer}: Rachel Zhou \email{rzhou14@illinois.edu} [copyright holder]

Authors:
\itemize{
  \item Dave Zhao \email{netidhere@illinois.edu} [copyright holder]
}

}
