% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ihmm.R
\name{ihmm}
\alias{ihmm}
\title{Estimation and Inference for High Dimensional Median Analysis}
\usage{
ihmm(Y, G, S, mediation_setting = "incomplete", tuning_method = "aic",
  lam_list = NA, min.ratio = 0.1, n.lambda = 5, center = TRUE)
}
\arguments{
\item{love}{Do you love cats? Defaults to TRUE.}
}
\description{
Amazing statistical computations are being undertaken.
}
\examples{
cat_function()
}
