% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LM_function.R
\name{lm_function}
\alias{lm_function}
\title{lm function}
\usage{
lm_function(carapace_length, body_depth)
}
\arguments{
\item{numarical}{data}
}
\value{
carapace length body_depth lm plot
}
\description{
lm function
}
