% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ihmm.R
\name{ihmm}
\alias{ihmm}
\title{Estimation and Inference for High Dimensional Median Analysis}
\usage{
ihmm(Y, G, S, mediation_setting = "incomplete", tuning_method = "aic",
  lam_list = NA, min.ratio = 0.1, n.lambda = 5, center = TRUE)
}
\arguments{
\item{Y}{The n-dimensional outcome vector.}

\item{G}{The n by p mediator matrix. p can be larger than n.}

\item{S}{The n by q exposure matrix. q can be 1, and q < n is required.}

\item{mediation_setting}{Place your description here...}

\item{tuning_method}{Place your description here...}

\item{lam_list}{Place your description here...}

\item{min.ratio}{Place your description here...}

\item{n.lambda}{Place your description here...}

\item{center}{Place your description here...}
}
\description{
Amazing statistical computations are being undertaken.
}
\examples{
## Example to be added here

# c.f. test_data.R file
}
