% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ihmm.R
\name{ihmm}
\alias{ihmm}
\title{A Cat Function}
\usage{
ihmm(Y, G, S, mediation_setting = "incomplete", tuning_method = "aic",
  lam_list = NA, min.ratio = 0.1, n.lambda = 5, center = TRUE)
}
\arguments{
\item{love}{Do you love cats? Defaults to TRUE.}
}
\description{
This function allows you to express your love of cats.
}
\examples{
cat_function()
}
\keyword{cats}
